zuzikoko/requirements.txt
acme==2.1.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
Babel==2.11.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
Brotli==1.0.9
bs4==0.0.1
celery==5.2.7
certbot==2.1.0
certbot-nginx==2.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
cryptography==38.0.3
cssselect2==0.7.0
cssutils==2.6.0
daphne==4.0.0
Deprecated==1.2.13
distlib==0.3.6
distro==1.8.0
dj-database-url==1.0.0
Django==4.1.2
django-crispy-forms==1.14.0
django-currencies==0.10.1
django-debug-toolbar==3.7.0
django-inlinecss==0.3.0
django-jet-4==0.0.1
django-localflavor==3.1
django-modeltranslation==0.18.5
django-money==3.0.0
django-pandas==0.6.6
django-parler==2.3
django-pytest==0.2.0
django-redis==5.2.0
django-rosetta==0.9.8
django-sslserver==0.22
django-translation-flags==1.0.6
docker==6.0.0
filelock==3.8.0
flower==1.2.0
fonttools==4.37.4
forex-python==1.8
future==0.18.3
glib==1.0.0
gobject==0.1.0
gunicorn==20.1.0
html5lib==1.1
humanize==4.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
joblib==1.2.0
josepy==1.13.0
jsonfield==2.1.1
kombu==5.2.4
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parsedatetime==2.6
Pillow==9.2.0
platformdirs==2.5.3
polib==1.1.1
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psycopg2==2.9.5
psycopg2-binary==2.9.4
py-moneyed==2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydyf==0.3.0
pynliner==0.8.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyphen==0.13.0
pyRFC3339==1.1
python-dateutil==2.8.2
python-stdnum==1.18
pytz==2022.2.1
rabbitmq==0.2.0
redis==4.3.4
requests==2.28.1
scikit-learn==1.2.0
scipy==1.10.0
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
sqlparse==0.4.3
stripe==4.2.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.2
translations==0.1.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
uWSGI==2.0.21
vine==5.0.0
virtualenv==20.16.6
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
websocket-client==1.4.1
wrapt==1.14.1
zope.interface==5.5.1
zopfli==0.2.1
